Wait for offsets initialization in the control plane
